requirements.txt
antlr4-python3-runtime==4.13.0
pyqt5==5.15.11